chore: move filesystem helpers to internal
